New WGEN system
I've deactivated the smooth populator until all biomes are properly
done.
This system should work for all standard biomes, simply change the
scale of the master noise, and pass it to the super class through the
constructor. More params can also be set to alter the look. To change
the ground cover, override the replace block method, and alter the
result of the super call.

Signed-off-by: Aleksi Sapon <QcTechs@gmail.com>