Begin moving everything over to protocol events
I got somewhat bored of this partway through, and there's a ton left to
do. Because I didn't want to have protocol event classes for most of the
packets used, I reused Vanilla packets as protocol events and just
returned the packet object in VanillaNetworkSynchronizer.
This is a bit of a cheat, but will work for now. What I've been thinking
about doing is making the data classes (notes, tile entities, etc)
implement ProtocolEvent and have the NetworkSynchronizer get the
appropriate packets to update the client of the data object's state.