Instead, lets have a VanillaEntityComponent that can hold sub components.
For example:
Enderman is a Component composed of EndermanModel, Health, EndermanAttack,
Wander, StepStound, and anymore that comes up. When spawning an entity
that we want to be an Enderman, just attach the Enderman component. It
also has a link to the protocol that Vanilla needs.

To determine if an entity is an Enderman,
entity.hasComponent(Enderman.class) is the method for now.

Signed-off-by: Chris Sanders <zidane@outlook.com>