Added /plugins-tofile.
This command outputs a file having informations about all loaded
plugins. The report is stored in plugins/Vanilla/pluginreports. The
commadsource can define the filename of the report but if no filename
was passed up on execution, then the current date and time is taken as
the filename for the report.