Update for API changes to world.
Removed the useless storage of Sky and the worlds it is attached to (it is
a WorldComponent, owner is all that is needed). Cleaned up a lot of other
sillyness that I saw.

@DDoSQc Since I removed height from world, I stuck the value in
VanillaNetworkSynchronizer. You may want to store it in the world's
datatable or somewhere else you see fit.
Signed-off-by: Chris Sanders <zidane@outlook.com>