Add and partially re-implement entity spawning items
This commit combines all known Vanilla items that can spawn entities
upon interaction under one single Class that deals with this. It
can be further extended with onSpawn or similar methods so item materials
can better specify the properties of the spawned Entity. For now, there
is only an onThrown for throwable items.

Signed-off-by: Irmo van den Berge <bergerkiller@gmail.com>